requirements.txt
pandas
python-dateutil
streamlit
openai
python-dotenv